+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## UNRELEASED
-- \[new feature\] Add `encryptStream.destroy(error)` and `decryptStream.destroy(error)` functions.
+- \[enhancement\] Correctly handle `encryptStream.destroy(error)` and `decryptStream.destroy(error)` functions.
 - \[deprecation\] Deprecate `encryptStream.emit('cancel')'` and `decryptStream.emit('cancel')` in favor of `stream.destroy()`
 
 ## v1.3.0 - 2025/03/20