+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## UNRELEASED
+## v1.4.0 - UNRELEASED
 - \[enhancement\] Correctly handle `encryptStream.destroy(error)` and `decryptStream.destroy(error)` functions.
 - \[deprecation\] Deprecate `encryptStream.emit('cancel')'` and `decryptStream.emit('cancel')` in favor of `stream.destroy()`
+- \[enhancement\] Update dependencies
+
 
 ## v1.3.0 - 2025/03/20
 - \[new feature\] Expose keySize for AsymKeys